Hacked in some very basic reels

+++ UnityProjects/LayoutEditor/Assets/_Project/Scripts/Oasis/Oasis.MAME/MameController.cs
@@ -14,16 +14,48 @@ namespace Oasis.MAME
     // can be same pattern as the ExtractIMporter, standrad C# class
     public class MameController : MonoBehaviour
     {
+        public bool ArgsOutputConsole;
+        public bool ArgsVideoNone;
+
+        public bool DebugOutputStdOut;
+
+
         private const string kTEMPHardcodedMameExeDirectoryPath = "Emulators\\MAME\\mame0258";
         private const string kTEMPHardcodedRomName = "j6popoli";
 
         private const string kMameExeFilename = "mame.exe";
 
+        // Outputs component state changes to console, may be better way of pulling these out of MAME
+        private const string kArgsOutputConsole = "-output console";
+
         // This '-video none' option means we don't need to actually skip the 'this game is not working' screens
         // and go straight into the emulation.  Also don't need -window or -skip_gameinfo either
-        private const string kAdditionalArgs = "-output console -video none -seconds_to_run 999999999";
+        private const string kArgsVideoNone = "-video none -seconds_to_run 999999999";
+        private const string kArgsForTestingWithVideo = "-window -skip_gameinfo";
 
-        private const string kLampDataPrefix = "lamp";
+        // "-state A";  this works, loads the required state as part of startup, very fast/clean
+
+        // -autosave no good, only works when save states explicitally supported, will need to 
+        // 'manually' save state on exit via sending Lua command.  Then can specify '-state 1' etc
+        // to load with a saved state
+
+
+        // Some sample MAME output data until implemented:
+        //digit2 = 23387  (JP prob 1 bit per segment, support up to 16 segs including dot?
+        //digit3 = 16191
+        //digit5 = 14392
+        //STATLED = 1
+        //reel1 = 94  JP Not sure whether should be using reel1 or sreel1 value, I think one of them is legacy for the old reel system in the MAME internal layouts
+        //sreel1 = 64170
+        //vfdduty0 = 27   (JP 0-31 I believe)
+        //vfd13 = 41164  (JP prob 1 bit per segment, support up to 16 segs including dot?  or is it for char set?
+
+        private const string kDataPrefixLamp = "lamp";
+        private const string kDataPrefixDigit = "digit";
+        private const string kDataPrefixReel = "reel";
+        private const string kDataPrefixVfdDuty0 = "vfdduty0";
+        private const string kDataPrefixVfd = "vfd";
+
 
 
         // new test approach to deal with callback blocking
@@ -31,7 +63,13 @@ namespace Oasis.MAME
         {
             get;
             private set;
-        } = new int[1024]; // TEMP test, no idea how large this needs to be!
+        } = new int[1024]; // TEMP test, no idea how large this needs to be wrt all techs!
+
+        public int[] ReelValues
+        {
+            get;
+            private set;
+        } = new int[16]; // TEMP test, no idea how large this needs to be wrt all techs!
 
 
         public UnityEvent OnImportComplete = new UnityEvent();
@@ -51,17 +89,34 @@ namespace Oasis.MAME
         private void OnDestroy()
         {
             // TODO this stuff will want to be in the StopMame() flow etc
+            if(_process != null)
+            {
+                _process.OutputDataReceived -= OnOutputDataReceived;
 
-            _process.OutputDataReceived -= OnOutputDataReceived;
+                _process.CancelOutputRead();
 
-            _process.CancelOutputRead();
-
-            _process.Kill();
+                _process.Kill();
+            }
         }
 
         public void StartMame()
         {
-            string arguments = kTEMPHardcodedRomName + " " + kAdditionalArgs;
+            string additionalArgs = "";
+            if(ArgsOutputConsole)
+            {
+                additionalArgs += " " + kArgsOutputConsole;
+            }
+
+            if (ArgsVideoNone)
+            {
+                additionalArgs += " " + kArgsVideoNone;
+            }
+            else
+            {
+                additionalArgs += " " + kArgsForTestingWithVideo;
+            }
+
+            string arguments = kTEMPHardcodedRomName + additionalArgs;
             _process = StartProcess(MameExeDirectoryFullPath, kMameExeFilename, arguments);
         }
 
@@ -108,16 +163,26 @@ namespace Oasis.MAME
 
         private void ProcessLine(string lineData)
         {
+            if(DebugOutputStdOut)
+            {
+                UnityEngine.Debug.LogError(lineData);
+            }
+
             // TODO very crude for now, will be able to be optimised with dictionaries etc
-            if(lineData.Substring(0, kLampDataPrefix.Length) == kLampDataPrefix)
+            if (lineData.Substring(0, kDataPrefixLamp.Length) == kDataPrefixLamp)
             {
                 ProcessLineLamp(lineData);
             }
+            else if (lineData.Substring(0, kDataPrefixReel.Length) == kDataPrefixReel)
+            {
+                ProcessLineReel(lineData);
+            }
         }
 
+        // TOIMPROVE - can make a generic function since this component number/value text extraction is copy/paste
         private void ProcessLineLamp(string lineData)
         {
-            int lampNumberStartIndex = kLampDataPrefix.Length;
+            int lampNumberStartIndex = kDataPrefixLamp.Length;
             int lampNumberEndIndex = lineData.IndexOf(' ');
             string lampNumberString = lineData.Substring(lampNumberStartIndex, lampNumberEndIndex - lampNumberStartIndex);
             int lampNumber = int.Parse(lampNumberString);
@@ -126,11 +191,27 @@ namespace Oasis.MAME
             string lampValueString = lineData.Substring(lampValueStartIndex, lineData.Length - lampValueStartIndex);
             int lampValue = int.Parse(lampValueString);
 
-            UnityEngine.Debug.LogError("lampNumber " + lampNumber + "   lampValue " + lampValue);
+            //UnityEngine.Debug.LogError("lampNumber " + lampNumber + "   lampValue " + lampValue);
 
             //OnLampChanged?.Invoke(lampNumber, lampValue);
 
             LampValues[lampNumber] = lampValue;
+        }
+
+        private void ProcessLineReel(string lineData)
+        {
+            int reelNumberStartIndex = kDataPrefixReel.Length;
+            int reelNumberEndIndex = lineData.IndexOf(' ');
+            string reelNumberString = lineData.Substring(reelNumberStartIndex, reelNumberEndIndex - reelNumberStartIndex);
+            int reelNumber = int.Parse(reelNumberString);
+
+            int reelValueStartIndex = lineData.LastIndexOf(' ');
+            string reelValueString = lineData.Substring(reelValueStartIndex, lineData.Length - reelValueStartIndex);
+            int reelValue = int.Parse(reelValueString);
+
+            //UnityEngine.Debug.LogError("reelNumber " + reelNumber + "   reelValue " + reelValue);
+
+            ReelValues[reelNumber] = reelValue;
         }
 
         private void OnOutputDataReceived(object sender, DataReceivedEventArgs dataReceivedEventArgs)
